privacy: use server id
